requirements.txt
earthengine-api
pandas
flask
folium
markdown
plotly
matplotlib
scipy

PyYAML
seaborn
fpdf2
pypdf
shapely